Fix broken integration with Prettier 3
Prettier 3 changes the way it formats some HTML, and one
of those changes caused any style block (CSS) code to re-indent on every
pass of prettier. (That's why I added tests for idempotence.) The new
placeholders used in style blocks solve this problem by not using
comments in places where there should be an actual value of some kind.
That makes things slightly more complex when the tokens are created, but
FAR simpler when they are replaced again.